Adds basic tests for new privatekey and cert options.
