+++ testdata/test-cert.crt
@@ -1,0 +1,1 @@
+this file is a pretend certificate for tests
